chore: refactor access to collections & type them directly
