+++ server/src/common/collections/schemas.js
@@ -1,0 +1,5 @@
+module.exports = {
+  logs: require("./logs"),
+  inserJeunesNationals: require("./inserJeunesNationals"),
+  formationsStats: require("./formationsStats"),
+};
